+++ jdbc/src/ConfigDataSource.kt
@@ -1,18 +1,24 @@
 package klite.jdbc
 
 import klite.Config
+import klite.info
+import klite.logger
 import java.io.PrintWriter
 import java.sql.Connection
 import java.sql.DriverManager
 import java.sql.SQLFeatureNotSupportedException
 import javax.sql.DataSource
 
-/** Non-pooled basic datasource */
+/** Non-pooled basic datasource, see also [PooledDataSource] */
 open class ConfigDataSource(
   val url: String = Config["DB_URL"],
   val user: String? = Config.optional("DB_USER"),
   private val pass: String? = Config.optional("DB_PASS"),
 ): DataSource {
+  init {
+    logger().info("Connecting to $url${user?.let { ", user: $user" }}")
+  }
+
   override fun getConnection() = getConnection(user, pass)
   override fun getConnection(user: String?, pass: String?): Connection = DriverManager.getConnection(url, user, pass)