+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Unreleased
 * server: TypedHttpClient/JsonHttpClient uses a better logger name (from nearest user class)
+* json/jackson: TypedHttpClient/JsonHttpClient now have overridable trimToLog property in case you need to process how requests/responses are logged
 * jobs: JobRunner.schedule() now takes kotlin.time.Duration values
-* jdbc: PooledDataSource - a simple and easy to configure connection pool (used by DBModule by default)
+* jdbc: PooledDataSource - a simple and easy to configure connection pool (used in DBModule by default)
 * jdbc: HikariModule is provided if you want to keep using it (add dependency on HikariCP manually)
 
 # 1.5.2