+++ CHANGELOG.md
@@ -1,9 +1,10 @@
-# Unreleased
+# 1.5.3
 * server: TypedHttpClient/JsonHttpClient uses a better logger name (from nearest user class)
 * json/jackson: TypedHttpClient/JsonHttpClient now have overridable trimToLog property in case you need to process how requests/responses are logged
 * jobs: JobRunner.schedule() now takes kotlin.time.Duration values
 * jdbc: PooledDataSource - a simple and easy to configure connection pool (used in DBModule by default)
 * jdbc: use HikariModule instead of DBModule if you still want to use Hikari (also add dependency on com.zaxxer:HikariCP)
+* liquibase/serialization - dependencies updated
 
 # 1.5.2
 * core: experimental Decimal class to be used for monetary values, with numerical equality (unlike BigDecimal)